pyproject.toml
[project]
name = "reddit-analytics-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.10.1",
    "pandas>=3.0.1",
    "praw>=7.8.1",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
